general fix

+++ app/src/main/java/com/ferenc/pamp/presentation/screens/main/chat/create_order/create_order_pop_up/CreateOrderPopUpPresenter.java
@@ -32,7 +32,7 @@ public class CreateOrderPopUpPresenter implements CreateOrderPopUpContract.Prese
     private boolean mHasOrder;
     private boolean mIsSendOrderListFlow;
     private GoodDealResponse mGoodDealResponse;
-    private int mSendOrderListQuantity;
+    private double mSendOrderListQuantity;
     private String mCleanPrice;
     private String mHonorar;
     private String mTotalPrice;
@@ -42,7 +42,7 @@ public class CreateOrderPopUpPresenter implements CreateOrderPopUpContract.Prese
                                      GoodDealResponseManager _goodDealResponseManager,
                                      CreateOrderPopUpContract.OrderModel _orderModel,
                                      boolean _isSendOrderListFlow,
-                                     int _sendOrderListQuantity) {
+                                     double _sendOrderListQuantity) {
         this.mView = _view;
         this.mOrderModel = _orderModel;
         this.mGoodDealResponseManager = _goodDealResponseManager;
@@ -75,7 +75,7 @@ public class CreateOrderPopUpPresenter implements CreateOrderPopUpContract.Prese
         }
     }
 
-    private void setLocalOrder(int _quantity) {
+    private void setLocalOrder(double _quantity) {
         mQuantity = _quantity;
         mView.showQuantity(String.valueOf(_quantity));
         calculateTotal();
